Make methods using IChannelConfiguration public and refactor topic calcs
IChannelConfigurations were implemented having extensibility in mind,
but in order to keep the APIs simple, were only used internally. Now,
publish and subscribe methods are heavily overloaded, and can receive
either IChannelConfigurations or tha parameters directly.

IChannelConfiguration has been redesigned to be back-end independent.
Now, the computation of the topic name (which is technology dependent)
is done in the Server class, rather than in the configuration class.