Add a cache to calculate topic patterns
This is done to optimize how it is checked that a given callback must be
executed when a new topic is received. This is because, to be sure that
the topics match, the actual parameters of a channel name must be
substituted.

Consequently to this change, all uses of IParameterLiteral[] have been
changed to List<IParameterLiteral> to make sure that "equals" and
"hashMap" work as expected when the IParameterLiterals are used in the
calculation of HashMaps (two java arrays are equals only when they are
the same object, and they are different even if their contents are
equals).